+++ src/Otros/Tree.java
@@ -48,7 +48,7 @@ public class Tree {
     public void moveDown(){
         if(trees.y<=2000) trees.y = (int) (trees.y + Velocity * InGame.delta_time);
 
-        if(new Rectangle(trees.x,trees.y,60,60).intersects(new Rectangle(Vehicle.vehiclePosition.x, Vehicle.vehiclePosition.y, Vehicle.W, Vehicle.H))){
+        if(new Rectangle(trees.x,trees.y,60,60).intersects(new Rectangle(Vehicle.vehiclePosition.x, Vehicle.vehiclePosition.y, (int)Vehicle.W, (int)Vehicle.H))){
 
             InGame.isPlaying = false;
